Allow docking text window into graphics window.
It can be convenient to dock the text window onto the graphics
window, e.g. for fullscreen support, for window managers that
don't handle two side-by-side windows easily, for ease of dragging,
etc. Also, emscripten only provides access to one OpenGL context,
so it's a requirement for a WebGL port.

This isn't yet fully usable: it misses scrollbar support, and
there is no way to resize the dock.