Updating dependencies in README
`libgtkmm` is wrong version is `x11` is missing from the packages to install. All git submodules were needed on my system (or at least more than just `libdxfrw`.)

Might also want to mention in the README that the repo should be cloned, not downloaded from the downloads page, but I'll leave that to someone else.