Added LoadFromResource to load ttf form resource
Added the public method LoadFromResource to load a TTF font from
a resource, i.e. from res/fonts/*.ttf.
To implement this, I moved the freetype2 processing code into its
own function, such that LoadFromFile and LoadFromResource only do
the loading and then hand off to the new method PostLoadProcess.

The lazy loading mechanism for fonts loaded from a file (nameOnly=
true) interfered with the resource loading, as LoadFont would call
LoadFromFile on a font actually provided by a resource.
To circumvent this, I removed the nameOnly option for resource fonts
such that they are fully loaded on initialisation of SolveSpace.
This solution is far from ideal (see my comment in the code), but
works as long as there are few resources.