Implemented Lazy Loading for Resource Fonts
This commit implements lazy loading for fonts given by resources.
This is achieved by tagging the fontFile of resource fonts with
the prefix 'res://' akin to protocol specifiers such as https://.
The LoadFont function then tests for 'res://' at the beginning of
a fontFile to decide whether to load the font from a file or
resource.

Tagging resource fonts in their file name seems like a better idea
than adding an is_resource boolean to the TtfFont class, as it will
cause assertions to throw more useful errors when a resource font
is loaded as a file and also ensures that no such file exists,
preventing unexpected behaviour.