Port IdList to using a shared_ptr<vector>.
The issue before is that much of the code relies on a copy
or assignment of IdList being a shallow copy,
so using a shared_ptr was the simplest way to preserve
that meaning. (Using a Vector directly was unsuccessful
due to changed copy semantics. The variety of iteration methods
makes a port to shared_ptr<unordered_map> harder and more far-reaching,
if nevertheless worthwhile.)