Prevent zero-length tangents and normals.
Lathed surfaces with a point on the lathe axis, there is a degenerate edge where one of the tangents is zero and the computed normal will be zero. That messes up booleans when the seams are in-plane with another surface. To prevent that we avoid taking tangents at the vertex and instead use on nearby.