Fix icon name for XDG mimetypes
The default icon for XDG mimetypes is the name of the mime with '/'
replaced with '-', unless the mimetype has an 'icon' attribute (and
solvespace-mime.xml does not have it). So the right icon name for
application/x-solvespace is 'application-x-solvespace', not
'application.x-solvespace'. Hence, properly rename the mimetype icons
in non-flatpak and non-snap installations.