A workplane can now be defined from a workplane defined by point and normal
It is now possible to create a "New Group | Sketch in New Workplane" from
an existing workplane defined using a point and a nowmal. Before we used to
hit the `ssassert(false, "Unexpected workplane subtype");`.

This makes 4308dc136b03d6727da96ec75cc7ae27d169040a more complete and
is related to #1120, #1048 and
https://github.com/solvespace/solvespace/pull/1054