Escape, AlternateTool() & right-click will now remove any incomplete curve in the process of being placed. It will also reselect the latest tool, but unselect the tool if no cure is in the process of being drawn.
Acomplished by using UndoUndo function which undoes the action registered after the operation is canceled, but does not add it to the redo stack so that Redo will not add the curve that was removed by the cancel operation.

CancelPending is the main function added to facilitate any trigure that cancels the current action in process. UndoUndo was altered to not store the undone action if false is passed in as an aurgument.

AlternateTool was also updated to work more efficently with CancelPending.

Last a specific command was added to the  MenuEdit function for the escape key.